pyproject.toml
[project]
name = "agent3dify"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "neka-nat", email = "nekanat.stock@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "cairosvg>=2.9.0",
    "deepagents>=0.4.12",
    "langchain>=1.2.15",
    "langchain-openai>=1.1.12",
    "numpy>=2.4.4",
    "pillow>=12.2.0",
    "python-dotenv>=1.2.2",
]

[project.scripts]
agent3dify = "agent3dify.cli:main"

[build-system]
requires = ["uv_build>=0.8.15,<0.9.0"]
build-backend = "uv_build"
